refactor: corrects misspell
